Update patch set 27
Patch Set 27:

Hey Anthony, Pete, and Felipe:

Thanks for your feedback.  I've tried to address some of the essential comments while improving the gate, and getting it back online.

Sorry for not responding to each point inline.

I think this is ready to merge, so we can start work on a proper API integration with Drydock.

Patch-set: 27