Update patch set 1
Patch Set 1:

> why not just put the content of requirements-direct.txt into
 > requirements.txt?  Or perhaps change the Dockerfile to do `pip
 > install -r requirements.txt`?

Hey Tin,

This is an area, where I feel python is behind modern dependency management practices.  I much prefer to have a "lock file" style of approach, where there is an automated method to update that lock file and only direct dependencies are specified.

This lets you update dependencies in a controlled way, while keeping the developer-specified/direct dependencies user friendly and meaningful.

Then intention of this approach is to do just that.  There is a tox -e freeze job that updates the requirements-frozen.txt file, and that is the only thing that should be installed inside the docker image.

I believe the names for these files in the various UCP projects are not yet standardized, but this approach is one we have generally agreed to.

If we just do the equivalent of installing requirements-direct.txt, then we cannot specify the versions of our indirect dependencies, which is essential for stability.

If we include our indirect dependencies in one place, it becomes a bit of a mess to manage, and it's easy to leave behind debris as code changes (especially an issue early in a project's life-cycle).

I'm definitely happy to discuss more on- or offline.  This is essentially the same approach with golang, rust, npm, and even perl (if using carton).

Patch-set: 1