Update patch set 2
Patch Set 2:

> I thought the  Armada wait was a feature that we could drive via
 > arguments
 > I would think the fix here would be to use it or not as an option

Armada has had "wait" functionality for some time, but now it waits in a different way that causes it to never think that the charts deployed here are ready.

We are asking for additional configuration for Armada to let us specify, e.g. some label selectors for how to decide whether a chart is done deploying (we will want something more conservative than the OSH release_group pattern for some charts, I think).

Additionally, Anthony and Hassan are working to update the existing charts so that they are generally in line with OSH conventions, which will make it compatible with the new waiting functionality.

Patch-set: 2
Reviewer: Mark Burnett <1011672@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0