Update patch set 17
Patch Set 17:

While this lays down the foundation for a Promenade API, it should still be tested in a follow-up PS.

I created this PS in DH just to demonstrate that it's straightforward: https://review.gerrithub.io/#/c/385479/

Ironically, in doing so, I fixed a bug. Which highlights why testing is so important.

Patch-set: 17
Label: Code-Review=0