Fix: Update use of cmp to expect correct exit code
This update to the usage of `cmp` in the anchor scripts will cause the
scripts to overwrite existing manifests whenever differences are
detected.

This should only occur during genesis, but may occur if a file is
accidentily modified on a node and the anchor then checks the file for
changes.

Change-Id: Ic594fef2c5a5c30a5be3ccc7ab51efcd29c8cdcc