Update patch set 1
Abandoned

Abandoning this.

Scott took a more complete stab at this, and I pushed on it a bit and got to a point that it needs a little deeper digging to get going.

We'll pickup Scott's change and carry that forward when we can: https://review.gerrithub.io/#/c/394615/

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon