Update patch set 1
Patch Set 1:

(2 comments)

Immediate issue appears to be that certs aren't getting inlined properly in the calico chart.

Using the default values of the chart (so endpoints...hosts.default: 10.96.232.136) it looks like the init container tries to wait for the serivce "kube-system/etcd", which is slightly unexpected.

Patch-set: 1
Label: Code-Review=0