Update patch set 1
Patch Set 1:

Hey Sam,

This looks like a great start.  Can you update the health check and join-script api calls in the resiliency gate to use this?  I guess that will require exposing some commands via the cli, but it will be good to have functional tests for this at a minimum.

Patch-set: 1
Label: Workflow=0