Update patch set 8
Patch Set 8:

> Hey Sam,
 > 
 > This is looking pretty good.  The big thing I would really like to
 > see is the use of this integrated into the resiliency gate.
 > 
 > Right now we do API calls like validatedesign and join-scripts with
 > curl directly, and it would be great to replace those.
 > 
 > I think most of that action is in tools/g2/stages/join-nodes.sh,
 > but it's possible there are one or two other corners in
 > tools/g2/lib/promenade.sh.
 > 
 > Best,
 > Mark

I don't have an opinion here but if the acceptance criteria for this story don't include resiliency gate integration, then can't this be merged and that integration be taken up in a separate patch set? Unless of course you'd like to validate that this works prior to merging in which case I agree that integration work is vital for acceptance.

Patch-set: 8